# services/backend-go/api/openapi.yaml
openapi: 3.1.0
info:
  title: SkoLab API
  version: "1.0"
  description: |
    The public API of skolab-api: the Go gateway. (The Python service runs in
    the same container and is not reachable from outside; compiles reach it
    through POST /api/v1/colab/compile.)

    Every error is `{"error": "<message for people>", "code": "<stable code>"}`;
    clients branch on `code`. Every response carries HSTS, nosniff and frame
    and referrer policies, and responses under /api are never cached, except
    the template catalog, which a browser may keep but must revalidate (ETag).

    The collaboration WebSocket (`GET /ws/colab/{workspace_id}?ticket=...`)
    is not described here: open it with a ticket from
    `POST /api/v1/ws/colab/{workspace_id}/tickets`.

    This file is the contract the QA tools test against (services/qa).
servers:
  - url: https://skolab-api.onrender.com
security:
  - firebase: []
...
paths:
...
  /hooks/sentry:
    post:
      tags: [health]
      operationId: sentryAlertHook
      summary: Sentry alert webhook, relayed to Slack (title and issue link only)
      description: >-
        Called by a Sentry internal integration, not by apps. Authenticated by
        Sentry-Hook-Signature, the hex HMAC-SHA256 of the raw body under the
        integration's client secret. Alerts (Sentry-Hook-Resource event_alert or
        metric_alert) are posted to Slack; other resources are acknowledged.
      security:
        - sentrySignature: []
      parameters:
        - name: Sentry-Hook-Resource
          in: header
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema: { type: object }
      responses:
        "204": { description: Verified; relayed to Slack or nothing to relay. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "401":
          description: Missing or invalid Sentry-Hook-Signature.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "429": { $ref: "#/components/responses/TooMany" }
        "502":
          description: Slack did not accept the alert.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /api/v1/users/profile/sync:
    post:
      tags: [users]
      operationId: syncProfile
      summary: Create or update the caller's own profile
      description: Required once per account before it can own or join a workspace.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ProfileSync" }
      responses:
        "200":
          description: Profile saved.
          content:
            application/json:
              schema:
                type: object
                required: [status, uid]
                properties:
                  status: { const: synced }
                  uid: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "429": { $ref: "#/components/responses/TooMany" }
        "500": { $ref: "#/components/responses/Internal" }
        "503": { $ref: "#/components/responses/Unavailable" }
  /api/v1/users/{userId}:
    delete:
      tags: [users]
      operationId: deleteAccount
      summary: Delete the caller's own account (data and sign-in)
      parameters:
        - { name: userId, in: path, required: true, schema: { type: string } }
      responses:
        "204": { description: Account deleted. }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409":
          description: >-
            owns_shared_workspaces. The caller owns workspaces other people
            actively collaborate in; deleting the account would delete them.
            Transfer each (POST /api/v1/workspaces/{id}/owner) or delete it,
            then retry. Nothing was deleted.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SharedWorkspacesConflict" }
        "429": { $ref: "#/components/responses/TooMany" }
        "500": { $ref: "#/components/responses/Internal" }
        "502":
          description: Data deleted, but removing the sign-in failed; retry.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/Unavailable" }
...
  /api/v1/ws/colab/{workspace_id}/tickets:
    parameters:
      - { name: workspace_id, in: path, required: true, schema: { type: string } }
    post:
      tags: [collaboration]
      operationId: issueTicket
      summary: A single-use, one-minute ticket for the collaboration WebSocket
      responses:
        "201":
          description: Ticket issued.
          content:
            application/json:
              schema:
                type: object
                required: [ticket, expires_in_seconds]
                properties:
                  ticket: { type: string }
                  expires_in_seconds: { type: integer, minimum: 1 }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "429": { $ref: "#/components/responses/TooMany" }
        "503": { $ref: "#/components/responses/Unavailable" }
...
  /api/v1/colab/compile:
    post:
      tags: [collaboration]
      operationId: compile
      summary: Compile LaTeX to PDF (charges the caller's quota; one at a time per user)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [latex_source]
              properties:
                latex_source: { type: string, minLength: 1, maxLength: 100000 }
                engine: { type: string, enum: [pdflatex] }
      responses:
        "200":
          description: The compile ran; `status` says whether it produced a PDF.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/CompileResult" }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "413": { $ref: "#/components/responses/TooLarge" }
        "429": { $ref: "#/components/responses/TooMany" }
        "502":
          description: The compile backend failed.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
        "503": { $ref: "#/components/responses/Unavailable" }
components:
  securitySchemes:
    firebase:
      type: http
      scheme: bearer
      bearerFormat: Firebase ID token
    sentrySignature:
      type: apiKey
      in: header
      name: Sentry-Hook-Signature
...
  schemas:
    Error:
      type: object
      required: [error, code]
      additionalProperties: false
      properties:
        error: { type: string, description: For people; wording may change. }
        code: { type: string, pattern: "^[a-z_]+$", description: Stable; branch on this. }
    SharedWorkspacesConflict:
      type: object
      required: [error, code, workspaces]
      additionalProperties: false
      properties:
        error: { type: string }
        code: { const: owns_shared_workspaces }
        workspaces:
          type: array
          minItems: 1
          maxItems: 100
          items:
            type: object
            required: [id, title, members]
            additionalProperties: false
            properties:
              id: { type: string }
              title: { type: string }
              members: { type: integer, minimum: 1, description: Active members besides the owner. }
...
    ProfileSync:
      type: object
      required: [uid, name]
      properties:
        uid: { type: string, minLength: 1, description: Must be the caller's own Firebase uid. }
        name: { $ref: "#/components/schemas/Text255" }
        discipline: { type: string }
    Text255:
      type: string
      description: 1-255 characters after trimming, no control characters.
      minLength: 1
      maxLength: 255
      pattern: "^[^\\u0000-\\u001f\\u007f-\\u009f\\s][^\\u0000-\\u001f\\u007f-\\u009f]*$"
...
    CompileResult:
      type: object
      required: [status]
      properties:
        status: { type: string, enum: [compiled, error, timeout] }
        pdf_base64: { type: [string, "null"] }
        log: { type: string }
        errors: { type: array, items: { type: string } }
...
  responses:
    BadRequest:
      description: The request is malformed or outside the contract.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unauthorized:
      description: Missing, invalid, expired or revoked Firebase ID token.
      headers:
        WWW-Authenticate: { schema: { type: string } }
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Forbidden:
      description: Signed in, but not allowed (role, account policy, or someone else's resource).
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    NotFound:
      description: Absent, or not visible to the caller (existence is never revealed).
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    TooLarge:
      description: Request body over 1 MiB (12 MiB for uploads and imports), or a project or file over its limit.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
...
    TooMany:
      description: Rate limited or quota spent; see Retry-After.
      headers:
        Retry-After: { schema: { type: integer } }
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Internal:
      description: Unexpected failure.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }
    Unavailable:
      description: A dependency (database, Firebase, compile backend) is unavailable.
      content: { application/json: { schema: { $ref: "#/components/schemas/Error" } } }